seperated login and register into routes
